requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
astroid==2.3.3
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.2
boto3==1.16.4
botocore==1.19.4
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
futures==3.1.1
goslate==1.5.1
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
multidict==4.7.6
numpy==1.17.3
oauthlib==3.1.0
pandas==0.25.3
PyDictionary==2.0.1
pylint==2.4.4
pyparsing==2.4.4
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
slack==0.0.2
slackclient==2.9.1
soupsieve==2.0.1
tweepy==3.9.0
typed-ast==1.4.0
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.0.29
Werkzeug==1.0.1
wrapt==1.11.2
yarl==1.6.0
zipp==3.1.0
